Fix Pointer cell typeIndex assignment

+++ LuxEngine/Core/Memory/Shared.hpp
@@ -50,11 +50,11 @@ namespace lux{
 
 
 		//Returns the CellClassIndex value of a CellClass
-		static constexpr inline uint32 classIndexFromEnum(const CellClass vClass) {
+		static constexpr inline uint16 classIndexFromEnum(const CellClass vClass) {
 			switch(vClass) {
 				#define _case(n) case CellClass::CLASS_##n: return (uint32)CellClassIndex::INDEX_##n;
 				_case(A) _case(B) _case(C) _case(D) _case(Q) _case(L)	//Fixed size classes
-				case CellClass::CLASS_0: return (uint32)-1;				//Custom size class
+				case CellClass::CLASS_0: return (uint16)-1;				//Custom size class
 				default: return -2; //This is just to suppress the warning
 			}
 		}
